Not needed
git-svn-id: svn://svn.code.sf.net/p/tass64/code/trunk@2431 111d7451-93fa-4e81-90d4-8039f76616e9

+++ my_getopt.c
@@ -208,7 +208,7 @@ static int my_getopt_internal(int argc, char *argv[], const char *shortopts,
           (argv[my_optind][charind] != '=');
         charind++);
     for (ind = 0; longopts[ind].name != NULL && hits == 0; ind++)
-      if ((strlen(longopts[ind].name) == (size_t) (charind - offset)) &&
+      if ((strlen(longopts[ind].name) == charind - offset) &&
          (strncmp(longopts[ind].name,
                   argv[my_optind] + offset, charind - offset) == 0))
         found = ind, hits++;
